Parser, calls, lets, closures
Add a parser that can recursively parse text from a StreamReader and
create the AST. Allow inner functions to be declared and function calls
to be made. Add let statements to declare a variable with a certain
type. Fix the process by which variables are bound to closures and
create a new closure for each function call.